Took another crack at CMake module structure -- may need to redo Project etc. on bulgogi side.
See, now that I've learned about hashmaps and have begun trying to use them on the CMake end of things, I'm realizing how much less convoluted/complicated it would be to convert the bulgogi project into a CMakeProject.
Ultimately, the internal program structs are meant to be manipulatable representations of the underlying project structures.
In this case, hashmaps in which the Target name is the key and the list of dependencies are the values seems best.
Also, each dependency can just be an enum which represents its own underlying type. There is now less to infer in the new architecture too.