Now extracting modules directly as HashMaps, removing duplicates along the way.
The extraction cleverly utilizes .dedup_by() inline to catch missed duplicates caused by the user inserting module1 and module1* as different entries in the project.
In the implementation, module1 and module1* would count as a duplicate, thus being ignored and not added to the list of modules.
Whichever appears first in the mapping (may be undefined ordering) will determine whether it is registered as an executable or normal module.
In the future, for shared library support to work, it may be a good idea to distinguish this differently.