Got somewhere with parsing the arguments and extracting their values...not sure why I am getting a segfault.
The segfault does not appear to make itself more clear with -fsanitize enabled.
HOWEVER it probably has something to do with the cli::client() function being outside the main() function.
The variables being used to store the arguments may be out of scope by the time the callback is run. Or something.