Added error handling to project.hpp just like client.hpp (see notes).
Except they are called with .err_if() and not just .err().
This is to distinguish the fact that every call to .err() _will_ complain because every enum member is an error.
However, the Project::Err is different in that it contains a `None` option. So, .err_if() takes this into account by doing nothing when e = Err::None.