Internal representation of Dependencies no longer lodges (*) in the name.
Instead, it is removed from the Dependency.name field and checked on Dependency::from() whether a * was conatained in the original project yaml.
This way, it can be stored as a bool Dependency.exe separately.
Conveniently, this happened to be true for both Targets and Modules. The need for a .exe that is, not the asterix notation.