Configured Dependency::from(Dependency::kind, string name) so that it would distinguish initializing targets from modules.
Specifically, this would affect how the .exe field is treated upon initialization.
By default, there is no way to determine the actually .exe property of a Target type Dependency just by the name.
This _is_ the case for modules.
So to address this, the Dependency::from() variant which does not take a bool exe specifier splits the kinds up.