Fixed that crazy stupid bug -- apparently map keys has special conditions.
Notes on these special conditions:
It **is** possible for a user-defined type to work, but it needs to conform to a set of special conditions.
Namely, its conversion to Key has to be possible, and I don't even want to get into that for now.
It would be good enough to have a lookup by target name and just be able to refer to them that way.