InDev v0.8.0 - The Inferencing Update.
Moved away from executable indicator primary determination to inferencing first and foremost.
When inferencing is not sufficient, an ambiguity error is displayed.
The user must intervene when necessary to resolve the ambiguity by using the executable indicator (*).

In most airtight cases it will not be needed, but projects earlier in development (with loose ends) may find themselves needing it more often.