Fixed that PESKY indirection issue.
It turns out that the LOCAL instances of the dep_list pointer did not subscribe to the change in the underlying address that a call to bul_engine_target_grow induced.
This meant that the local version actually stored an older, outdated version of the malloc()'d array.
When you try to realloc an old malloc pointer that was already realloc'd, that's a double free.
A double free in plain sight, right under your nose.
Thank god that was fixed.