C v0.1.10 - The C Engine Update.
It has been decreed from on high that the inference engine shall be written in C.
It is fast. It is simple. It is elegant. And most of all, free of bloat.
Upgrading to C v0.1.2 felt too insignificant, and C v0.2.0 felt premature.
C v0.1.10 declares something significant has happened, but no new features.