Added BUL_FS_WARN clause for filesystem errors.
Allows for non-fatal notices about perror and errno type returns.
In this case, when mkdir() returns -1 on the account of an EEXIST errno, it really should not be considered fatal behaviour.
The caller _should_ have the option, however, given that it could be valuable to even print warnings about these things.