Got the new core up until a certain point (see notes).
Most of the basis is good but I seem to have hit a wall.
Specifically, right around how to deal with 'unwinding' when a DOCUMENT_END or a SEQUENCE_END is encountered.
It is much easier to cap the level of instrospection to 1 because it means a single 'previous' variable can track the way back.

Current ideas revolve around a **parents array in core or having each target struct store its parent.