Level now increased on the spot (doc/scalar) + MAPPING_START effect deferred to next scalar.
Longer version: rather than wait for the next MAPPING_START event to advance (reveal) the stack value, it is both updated and leveled up at the same time.
The exception is MAPPING events which need to be deferred until the next scalar value.
This required the addition of a map flag to the core in which the scalar call 'catches' the flag when positive.
It must be reset every time it is found to be positive.