Using correct YAML syntax (see notes).
It turns out that the YAML document root is a single node.
Without the indication fo a sequence, the document can only support one entry.
It can be one mapping or one scalar. The only way it can be more is if it begins as a sequence.
This could be an interesting idea for root targets -- the document begin event may not need to induce a document target.